requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
anyio==3.7.0
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
chromadb==0.3.26
click==8.1.3
clickhouse-connect==0.6.4
coloredlogs==15.0.1
contourpy==1.1.0
cycler==0.11.0
dataclasses-json==0.5.8
decorator==5.1.1
duckdb==0.8.1
faiss-cpu==1.7.4
Faker==18.11.1
fastapi==0.98.0
favicon==0.7.0
filelock==3.12.2
flatbuffers==23.5.26
fonttools==4.40.0
frozenlist==1.3.3
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.31
h11==0.14.0
hnswlib==0.7.0
htbuilder==0.6.1
httptools==0.5.0
huggingface-hub==0.15.1
humanfriendly==10.0
idna==3.4
importlib-metadata==6.7.0
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
langchain==0.0.154
lxml==4.9.2
lz4==4.3.2
Markdown==3.4.3
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
mdurl==0.1.2
monotonic==1.6
more-itertools==9.1.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.25.0
onnxruntime==1.15.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
overrides==7.3.1
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
posthog==3.0.1
protobuf==3.20.3
pulsar-client==3.2.0
pyarrow==12.0.1
pydantic==1.10.9
pydeck==0.8.1b0
Pygments==2.15.1
pymdown-extensions==10.0.1
Pympler==1.0.1
pyparsing==3.1.0
pypdf==3.11.0
PyPDF2==3.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
rich==13.4.2
semver==3.0.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.17
st-annotated-text==4.0.0
starlette==0.27.0
streamlit==1.23.1
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.5
streamlit-embedcode==0.1.2
streamlit-extras==0.2.7
streamlit-faker==0.0.2
streamlit-image-coordinates==0.1.5
streamlit-keyup==0.2.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
sympy==1.12
tenacity==8.2.2
terminal==0.4.0
tiktoken==0.4.0
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.3
uvicorn==0.22.0
uvloop==0.17.0
validators==0.20.0
watchdog==3.0.0
watchfiles==0.19.0
websockets==11.0.3
yarl==1.9.2
zipp==3.15.0
zstandard==0.21.0
